@@ -17,41 +17,42 @@ import UserModel from "./user.model";
   tableName: "Companies",
   timestamps: false
 })
+
 export default class CompanyModel extends Model<CompanyModel> {
   @IsUUID(4)
   @PrimaryKey
   @Default(DataType.UUIDV4)
   @Column(DataType.UUID)
-  public id: string;
+  public id!: string;
 
   @Column
-  public name: string;
+  public name!: string;
 
   @AllowNull(true)
   @Column
-  public description: string;
+  public description!: string;
 
   @Column
-  public email: string;
+  public email!: string;
 
   @Column
-  public phone: string;
+  public phone!: string;
 
   @Column
-  public logo: string;
+  public logo!: string;
 
   @Column
-  public cover: string;
+  public cover!: string;
 
   @HasMany(() => UserModel)
-  public users: UserModel[];
+  public users!: UserModel[];
 
   @HasMany(() => ProductModel)
-  public products: ProductModel[];
+  public products!: ProductModel[];
 
   @Column
-  public createdAt: Date;
+  public createdAt!: Date;
 
   @Column
-  public updatedAt: Date;
+  public updatedAt!: Date;
 }
