@@ -4,16 +4,16 @@ import CompanyModel from "../../database/models/company.model";
 
 export default {
   Query: {
-    getAllCompanies: async (_parent, _args, _context) => {
+    getAllCompanies: async (_parent: any, _args: any, _context: any) => {
       return CompanyModel.findAll({
         include: [ProductModel, UserModel]
       });
     }
   },
   Mutation: {
-    createCompany: async (_parent, _args, { user }) => {
+    createCompany: async (_parent: any, _args: any, {user}: any) => {
       const newCompany = await CompanyModel.create({..._args}).then((company) =>  {
-        company.addUser(user.id);
+        company.$add("users", user.id);
         return company;
       });
       // CompanyModel.update(
