@@ -5,33 +5,32 @@ import CompanyModel from "../../database/models/company.model";
 
 export default {
   Query: {
-    getAllProducts: async (_parent, _args, _context) => {
+    getAllProducts: async (_parent: any, _args: any, _context: any) => {
       return ProductModel.findAll({
         include: [CategoryModel,CompanyModel]
       });
     },
-    getProduct: async (_parent, { id }) => {
-      return ProductModel.findOne({
-        id,
-        include: [CategoryModel, CompanyModel]
-      });
+    getProduct: async (_parent: any, { id }: any) => {
+      return ProductModel.findOne({where: {id: id}, include: [CategoryModel, CompanyModel]});
     }
   },
   Mutation: {
-    createProduct: async (_parent, _args, { user }) => {
+    createProduct: async (_parent: any, _args: any, {user}: any) => {
       let product = await ProductModel.create({ ..._args, companyId: user.companyId }).then((product) =>  {
         return product;
       });
       return product.toJSON();
     },
     addCategoryToProduct: async (
-      _parent,
-      { productId, categoryName },
-      _context
+      _parent: any,
+      { productId, categoryName}: any,
+      _context: any
     ) => {
       let category = await CategoryModel.findOne({
         where: { name: categoryName }
       });
+      if (!category)
+        return null;
       // findOrCreate so that it doesn't add multiple times the category to a product.
       await ProductCategoryModel.findOrCreate({
         where: {
@@ -40,7 +39,7 @@ export default {
         }
       });
       let product = await ProductModel.findOne({
-        id: productId,
+        where: {id: productId},
         include: [
           {
             model: CategoryModel,
