@@ -1,19 +1,21 @@
 import UserModel from "../../database/models/user.model";
-import { UserInputError } from "apollo-server-express";
 import { generateAuthlink } from "../../auth";
 import jwt from "jsonwebtoken";
-import { AuthenticationError, UserInputError } from 'apollo-server';
-const createToken = async (user, secret, expiresIn) => {
+import { AuthenticationError, UserInputError } from 'apollo-server-azure-functions';
+import {Context} from "@azure/functions"
+
+const createToken = async (user: any, secret: any, expiresIn: any) => {
   const { id, email, username, role } = user;
   return await jwt.sign({ id, email, username, role }, secret, {
     expiresIn,
   });
 };
 
-const EXPIREIN = '3000m'
+const EXPIREIN = '3000m';
 export default {
   Query: {
-    getUser: async (_parent, _args, { user }) => {
+    getUser: async (_parent: any, _args: any, { user }: any) => {
+      console.log(user);
       if (!user) {
         return null;
       }
@@ -22,7 +24,7 @@ export default {
     }
   },
   Mutation: {
-    login: async (_parent, {email, password}, {secret}) => {
+    login: async (_parent: any, {email, password}: any, {secret}: any) => {
       let user = await UserModel.findByLogin(email);
       if (!user) {
         throw new UserInputError(
@@ -36,7 +38,7 @@ export default {
       }
       return {token: createToken(user, secret, EXPIREIN), userId: user.id};
     },
-    register: async (_parent, {email, ...userInformations}, {secret}) => {
+    register: async (_parent: any, {email, ...userInformations}: any, {secret}: any) => {
       const emailAlreadyTaken = await UserModel.findOne({
         where: {email}
       });
@@ -49,7 +51,7 @@ export default {
 
       // TODO : send the generated url to the user via email.
       // this is the link generated by jwt to validate the user. Give it to the user by email to validate his account.
-      const validationLink = generateAuthlink("check-email", {
+      const validationLink = generateAuthlink("checkEmail", {
         id: user.id
       });
       console.log(validationLink);
@@ -57,7 +59,7 @@ export default {
       return {
         token: createToken(user, secret, "365 days"),
         userId: user.id,
-        message: `Un email de confirmation a été envoyé a cette adresse email : ${user.email}, clique sur le lien dans le mail afin valider ton compte !`
+        message: `Un email de confirmation a été envoyé a cette adresse email : ${user.email}, clique sur le lien ${validationLink} dans le mail afin valider ton compte !`
       };
     }
   }
