@@ -1,4 +1,4 @@
-import { gql } from "apollo-server";
+import { gql } from "apollo-server-azure-functions";
 
 export default gql`
   extend type Query {
