@@ -1,4 +1,4 @@
-import { gql } from "apollo-server-express";
+import { gql } from "apollo-server-azure-functions";
 
 import userSchema from "./user";
 import productSchema from "./product";
